Add files via upload
The eight queens puzzle is the problem of placing eight chess queens on an 8×8 chessboard 
so that a solution requires that no two queens share the same row, column, or diagonal. 
The eight queens puzzle is an example of the more general n queens problem of placing n non-attacking queens on an n×n chessboard, 
for which solutions exist for all natural numbers n with the exception of n = 2 and n = 3.